timestamp in seconds, fix force calibration

+++ SessionData.cs
@@ -8,7 +8,7 @@ namespace SpasticityClientV2
 {
     public class SessionData
     {
-        public long TimeStamp { get; set; }
+        public float TimeStamp { get; set; }
         public float Angle_deg { get; set; }
         public float AngVel_degpersec { get; set; }
         public float EMG_mV { get; set; }
